requirements.txt
fastapi
uvicorn
python-multipart
streamlit
ultralytics
requests
pillow
numpy